// src/enums/workspace.enum.ts
export enum Workspace {
    CURRENT_AWS_PROFILE = 'currentAwsProfile',
    USER_DATA = 'workspaceData',
    STAGE_SUPPORT = 'stageSupport',
    STAGE_LIST = 'stageList',
    CURRENT_STAGE = 'currentStage',
    PREFIX_NAME = 'prefixName',
    EXTENSION_CONFIGURED = 'isExtesionConfigured',
    AWS_CLI_COMMAND = 'awsCliCommand',
    SERVERLESS_CLI_COMMAND = 'awsServerlessCliCommand',
    TERMINAL_MODE = 'terminalMode',
    AWS_REGION = 'awsRegion',
    LOG_TIME = 'logTimeString'
}